broke out index into multiple compionents
